tx: consolidate generic tx capabilities
