+++ packages/tx/src/eip1559Transaction.ts
@@ -6,7 +6,6 @@ import {
   bytesToBigInt,
   bytesToHex,
   concatBytes,
-  ecrecover,
   equalsBytes,
   hexToBytes,
   toBytes,
@@ -15,6 +14,7 @@ import {
 import { keccak256 } from 'ethereum-cryptography/keccak.js'
 
 import { BaseTransaction } from './baseTransaction.js'
+import * as Generic from './capabilities/generic.js'
 import { TransactionType } from './types.js'
 import { AccessLists } from './util.js'
 
@@ -310,19 +310,7 @@ export class FeeMarketEIP1559Transaction extends BaseTransaction<TransactionType
    * Use {@link FeeMarketEIP1559Transaction.getMessageToSign} to get a tx hash for the purpose of signing.
    */
   public hash(): Uint8Array {
-    if (!this.isSigned()) {
-      const msg = this._errorMsg('Cannot call hash method if transaction is not signed')
-      throw new Error(msg)
-    }
-
-    if (Object.isFrozen(this)) {
-      if (!this.cache.hash) {
-        this.cache.hash = keccak256(this.serialize())
-      }
-      return this.cache.hash
-    }
-
-    return keccak256(this.serialize())
+    return Generic.hash.bind(this)()
   }
 
   /**
@@ -336,35 +324,7 @@ export class FeeMarketEIP1559Transaction extends BaseTransaction<TransactionType
    * Returns the public key of the sender
    */
   public getSenderPublicKey(): Uint8Array {
-    if (this.cache.senderPubKey !== undefined) {
-      return this.cache.senderPubKey
-    }
-
-    if (!this.isSigned()) {
-      const msg = this._errorMsg('Cannot call this method if transaction is not signed')
-      throw new Error(msg)
-    }
-
-    const msgHash = this.getMessageToVerifySignature()
-    const { v, r, s } = this
-
-    this._validateHighS()
-
-    try {
-      const sender = ecrecover(
-        msgHash,
-        v! + BigInt(27), // Recover the 27 which was stripped from ecsign
-        bigIntToUnpaddedBytes(r!),
-        bigIntToUnpaddedBytes(s!)
-      )
-      if (Object.isFrozen(this)) {
-        this.cache.senderPubKey = sender
-      }
-      return sender
-    } catch (e: any) {
-      const msg = this._errorMsg('Invalid Signature')
-      throw new Error(msg)
-    }
+    return Generic.getSenderPublicKey.bind(this)()
   }
 
   protected _processSignature(v: bigint, r: Uint8Array, s: Uint8Array) {
@@ -421,6 +381,6 @@ export class FeeMarketEIP1559Transaction extends BaseTransaction<TransactionType
    * @hidden
    */
   protected _errorMsg(msg: string) {
-    return `${msg} (${this.errorStr()})`
+    return Generic.errorMsg.bind(this)(msg)
   }
 }
