+++ packages/tx/src/legacyTransaction.ts
@@ -4,7 +4,6 @@ import {
   bigIntToHex,
   bigIntToUnpaddedBytes,
   bytesToBigInt,
-  ecrecover,
   toBytes,
   unpadBytes,
   validateNoLeadingZeroes,
@@ -12,6 +11,7 @@ import {
 import { keccak256 } from 'ethereum-cryptography/keccak.js'
 
 import { BaseTransaction } from './baseTransaction.js'
+import * as Generic from './capabilities/generic.js'
 import { Capability, TransactionType } from './types.js'
 
 import type {
@@ -256,19 +256,7 @@ export class LegacyTransaction extends BaseTransaction<TransactionType.Legacy> {
    * Use {@link Transaction.getMessageToSign} to get a tx hash for the purpose of signing.
    */
   hash(): Uint8Array {
-    if (!this.isSigned()) {
-      const msg = this._errorMsg('Cannot call hash method if transaction is not signed')
-      throw new Error(msg)
-    }
-
-    if (Object.isFrozen(this)) {
-      if (!this.cache.hash) {
-        this.cache.hash = keccak256(RLP.encode(this.raw()))
-      }
-      return this.cache.hash
-    }
-
-    return keccak256(RLP.encode(this.raw()))
+    return Generic.hash.bind(this)()
   }
 
   /**
@@ -286,32 +274,7 @@ export class LegacyTransaction extends BaseTransaction<TransactionType.Legacy> {
    * Returns the public key of the sender
    */
   getSenderPublicKey(): Uint8Array {
-    if (this.cache.senderPubKey !== undefined) {
-      return this.cache.senderPubKey
-    }
-
-    const msgHash = this.getMessageToVerifySignature()
-
-    const { v, r, s } = this
-
-    this._validateHighS()
-
-    try {
-      const sender = ecrecover(
-        msgHash,
-        v!,
-        bigIntToUnpaddedBytes(r!),
-        bigIntToUnpaddedBytes(s!),
-        this.supports(Capability.EIP155ReplayProtection) ? this.common.chainId() : undefined
-      )
-      if (Object.isFrozen(this)) {
-        this.cache.senderPubKey = sender
-      }
-      return sender
-    } catch (e: any) {
-      const msg = this._errorMsg('Invalid Signature')
-      throw new Error(msg)
-    }
+    return Generic.getSenderPublicKey.bind(this)()
   }
 
   /**
@@ -413,6 +376,6 @@ export class LegacyTransaction extends BaseTransaction<TransactionType.Legacy> {
    * @hidden
    */
   protected _errorMsg(msg: string) {
-    return `${msg} (${this.errorStr()})`
+    return Generic.errorMsg.bind(this)(msg)
   }
 }
